work on changing qrcodemodal to properly handle new qrcode image based on walletappId
